Refine blob storage interface
